Trying to merge scolopendra free bone animation with normal node animation
The most important thing is in Model::ReadNodeHierarchy.
Basically we got 2 GlobalTransformations of a bone - 1st is for global postition which we change, 2nd is for storing starting position of the bone.
In ReadNodeHierarchy we calculate globalPosition (parent * node) then, if bone is free we are removing from globalPosition translation from GlobalTransformation2 of the bone. Now we are meging global transformation with node animation. ALMOST. Only translations works, rotations are fucked up.